inference-code/requirements.txt
torch
diffusers
transformers
tqdm
Pillow
accelerate
